Created a Python Backend
Wanted to use python as the backend of finding coordinates of location because there is a package that allows for easy use. Had to set up a 3rd server being run on port 8000 in order for it to talk to the front end. Currently AboutPage.js is being used to test that.